@@ -9,22 +9,21 @@
     "eject": "react-scripts eject",
     "lint": "yarn run lint-prettier && yarn run lint-eslint",
     "lint-eslint": "eslint --ignore-path .gitignore src/**/*{ts,tsx} --fix",
-    "lint-prettier": "prettier --write \"src/**/*.(ts|tsx)\"",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public"
+    "lint-prettier": "prettier --write \"src/**/*.(ts|tsx)\""
   },
   "dependencies": {
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/styled-components": "^5.1.21",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/styled-components": "^5.1.25",
     "axios": "^0.26.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "styled-components": "^5.3.3",
+    "react-scripts": "5.0.1",
+    "styled-components": "^5.3.5",
+    "styled-reset": "^4.4.1",
     "typescript": "^4.5.5"
   },
   "eslintConfig": {
@@ -56,16 +55,6 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-interactions": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
-    "@storybook/builder-webpack5": "^6.4.22",
-    "@storybook/manager-webpack5": "^6.4.22",
-    "@storybook/node-logger": "^6.4.22",
-    "@storybook/preset-create-react-app": "^4.1.0",
-    "@storybook/react": "^6.4.22",
-    "@storybook/testing-library": "^0.0.11",
     "@typescript-eslint/eslint-plugin": "^5.16.0",
     "@typescript-eslint/parser": "^5.16.0",
     "eslint": "^7.11.0",
